Extract chart pull-to-refresh into ChartRefreshController
